still trying to fix initialization

+++ code/Util.h
@@ -7,7 +7,7 @@
 
 
 
-int getHammingDistance(gType v1, gType v2);
+unsigned long getHammingDistance(gType v1, gType v2);
 
 gType getRandomChild(gType p1, gType p2);
